backend/requirements.txt
celery==5.2.7
channels==4.0.0
channels-redis==4.1.0
daphne==4.0.0
Django==3.2
django-cors-headers==3.14.0
django-elasticsearch-dsl==7.3
django-elasticsearch-dsl-drf==0.22.5
django-environ==0.10.0
django-filter==23.1
django-mptt==0.14.0
django-nine==0.2.7
django-redis==5.2.0
django-silk==5.0.3
djangorestframework==3.14.0
djangorestframework-camel-case==1.2.0
djangorestframework-simplejwt==4.7.1
drf-nested-forms==1.1.7
drf-spectacular==0.26.1
drf-yasg==1.21.5
Pillow==9.4.0
redis==4.5.4
requests==2.30.0
